Fix '/' and stripTrailingSlash
When using the server setting stripTrailingSlash and serving and endpoint at '/',
hapi-api-version would search for v2 at '/v2/'. This route is impossible when
stripTrailingSlash is set to true in Hapi, so the fallback default route would be served.
This patch adds a new test showing the issue and a fix so that '/' will serve at '/v2'.